+++ app/src/main/java/org/koitharu/kotatsu/ui/reader/ReaderListener.kt
@@ -7,5 +7,5 @@ interface ReaderListener : BaseMvpView {
 
 	fun onPageChanged(chapter: MangaChapter, page: Int, total: Int)
 
-	fun saveState(chapterId: Long, page: Int, scroll: Float)
+	fun saveState(chapterId: Long, page: Int, scroll: Int)
 }
